moved options into Menu.h to be used in MiniMax
